Function References

Configuration Types

SpatialGSP.WarpConfig — Type

@kwdef struct WarpConfig{T <: AbstractFloat}
    N_bands::Int = 20
    λ_max::T = 2.0 # in (0, 2]
    iterate_threshold::T = convert(T, 10)
    discount_factor::T = convert(T, 0.5)
    max_iters::Int = 10
    aggregate_option::Symbol = :sum_normalized
    verbose::Bool = false
end

aggregate_option can be :sum_normalized or :sum

SpatialGSP.AxisSearchConfig — Type

@kwdef struct AxisSearchConfig{T <: AbstractFloat} <: GraphConstructionConfig
    kernel_σ_multiplier::T = convert(T, 3)
    make_undirected::Bool = true
    remove_null_nbs::Bool = true
    w_lb::T = convert(T, 0.1)
end

If remove_null_nbs is set to false, you can infer which neighbors came from which axis direction from the generated graph.

SpatialGSP.KNNConfig — Type

@kwdef struct KNNConfig{T <: AbstractFloat} <: GraphConstructionConfig
    single_distance::Bool = false
    atol::T = convert(T, 1e-3)
    normalize_min_weight::Bool = false
    make_symmetric::Bool = true
    k::Int = 4
    kernel_gain = convert(T, Inf)
    kernel_σ² = convert(T, Inf)
end

If normalize_min_wieght is false, then the minimum weight is 1.

Graph construction

SpatialGSP.create_axis_graph — Function

create_axis_graph(
    config::AxisSearchConfig{T},
    X::Vector{AV},
) where {T, AV <: AbstractVector{T}}

Constructs an undirected graph via the axis-search algorithm.

Warp sample construction

SpatialGSP.get_grid_warp_samples — Function

get_grid_warp_samples(
    x_nD::AbstractArray{T},
    config::WarpConfig,
) where T <: Real

Returns:

W, an array of the same size as x_nD that contains the warp samples.

G, the unit graph associated with x_nD.

SpatialGSP.get_knn_warp_samples — Function

get_knn_warp_samples(
    X::AbstractVector,
    y::AbstractVector{T},
    config::WarpConfig,
    knn_config::KNNConfig;
    check_finite::Bool = true,
) where T <: Real

To get an undirected G output, make sure knn_config.make_symmetric is true.

Returns:

W, warp samples for dataset (X,y).

G, the k-nearest neighbors graph for the dataset.

SpatialGSP.get_warp_samples — Function

get_warp_samples(
    x::Vector{T},
    TL::Union{Matrix, SparseMatrixCSC},
    config::WarpConfig,
) where T <: AbstractFloat

TL is the fundamental one-hop operator. One choice is the random-walk Laplacian of a graph.

x is the graph signal from which we derive the warp samples from.

Returns:

W, warp samples for dataset (X,y).

Graph matrices

SpatialGSP.create_adjacency — Function

create_adjacency(g::UnitGrid)

Returns the adjacency matrix of g.

create_adjacency(g::UndirectedGraph)

SpatialGSP.create_rwlaplacian — Function

create_rwlaplacian(::Type{T}, g::UnitGrid) where T

Returns the random-walk Laplacian of a unit grid graph, with scalar data type T.

create_rwlaplacian(g::UndirectedGraph)

SpatialGSP.create_snlaplacian — Function

create_snlaplacian(::Type{T}, g::UnitGrid) where T

Returns the symmetric normalized Laplacian of a unit grid graph, with scalar data type T.

create_snlaplacian(g::UndirectedGraph)

SpatialGSP.create_degree — Function

create_degree(g::UndirectedGraph)

Returns the degree matrix of graph g.

create_degree(g::UndirectedGraph)

SpatialGSP.create_laplacian — Function

create_laplacian(g::UndirectedGraph)

Returns the combinatorial Laplacian of g.

Utilities

SpatialGSP.convertcompactdomain — Function

convertcompactdomain(x::T, a::T, b::T, c::T, d::T)::T

converts compact domain x ∈ [a,b] to compact domain out ∈ [c,d].